# src/main/java/com/example/simpleboard/board/model/BoardDto.java
package com.example.simpleboard.board.model;

import com.example.simpleboard.post.entity.PostEntity;
import com.example.simpleboard.post.model.PostDto;
import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.*;

import java.util.List;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
@ToString
@JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
public class BoardDto {

    private Long id;

    private String boardName;

    private String status;

    @Builder.Default
    private List<PostDto> postList = List.of();

}